20190207_3

+++ spring_web_project/nxmile_approval/src/main/java/com/tpcom_apr/domain/service/MempntuptOutputVO.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MempntuptOutputVO {
-    private Long result_flag;
+    private int result_flag;
 }
